StreamIO: fix "bad write retry" in SSL mode
When a large frame is written to the wire over an SSL connection, PHP's OpenSSL extension behaves erroneously upon write failures in non-blocking mode. The fwrite call will return 0. However, if the variable containing the buffer is modified or reassigned, PHP will pass a different pointer to SSL_write(), causing the "bad write retry" error.

This behavior is also produced when a large frame (>= 49,152 bytes) is written directly to the wire with fwrite().

This patch resolves the issue by changing the following behavior:

  - Limits fwrite() to 8,192 bytes at a time
  - Does not reassign $data if 0 bytes are written