Update amqp_consumer_non_blocking.php
If we send a command to rabbitmq server while consuming a queue, client may receive an AmqpMessage before the command response. The message will be queued in this 'method_queue' and we will no longer detect we have something to read in 'stream_select', so we have to check if there is something queued in the channel and call 'wait' then.