Throw exception on missed heartbeat
If an IO detects missed heartbeats, it should not reconnect because
the AMQP connection and channels will not be recovered and server-side
connection and channels will be stopped on TCP connection close or
heartbeat timeout.
The current reconnection logic leads to delayed exception, if the client
waits for messages or tries to publish/ack.
It also will create a TCP connection on the server, which will consume
socket descriptors and memory.

Replace reconnect on missed heartbear with close and exception.

Since handling exceptions is a way to address connection issues in the
client, reconnection logic should be moved to exception handlers and be
higher-level, so channels and consumers can be recovered.