Introduce a method to create connection from multiple hosts.
Add a new static method `create_connection` in `AbstractConnection`
to create a connection providing multiple hosts.

The method takes an array of host arrays
with host,port,username,password and vhost keys
and tries to connect to them one-by-one.
The first successful connection is returned.
If unable to connect to any of them - the last esception is thrown.

This method is a simplest way to address issues like #253
There is no bookkeeping and load balancing, hosts are
selected one-by-one starting from the first one.

Combined with connection recovery techniques it can be used
to improve reliability when using clusters.